tyche-kvm: add capabilities.h
This commit copies over the vmx/capabilities.h to the tyche directory.
It seems like we just need to call into tyche and set the extern
variable `vmcs_config` and `vmx_capability` correctly inside
tyche/tyche.c, and the rest of the functions can be re-used in the exact
same manner without changes.

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>