tyche: move tyche driver inside kernel
Compile the tyche driver in-tree, so that we can export the symbol of
some of the tyche driver functions and call them from the kvm tyche
backend.

If we keep the tyche driver out of tree, then it would be impossible to
link these external tyche driver symbols and the linux kernel fails to
compile.

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>