tyche: expose raw_memory_add
Pulling @aghosn's changes on driver in commit 28b351913e391. I didn't
cherry-pick so that I can read the code a bit more carefully...

This commit adds the driver interface to allow bypass on the kvm driver,
so that the kvm-tyche can directly be exposed the non-continuous tyche
region for the domain. (Tyche is suppose to in charge of the EPT, not
KVM as we don't trust the VMM).

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>