tyche: vmx_vcpu_load
When we have a world switch on vmcs, vmx_vcpu_load calls the
vmx_vcpu_load_vmcs to do a vmcs clear and ptrld. This corresponds to a
tyche domain switch operation.

However, KVM calls the vmx_vcpu_load before completely setup the vmcs
fields of the vcpu. To do a tyche switch there, tyche is expecting a
handle on the domain, which requires sealing. A better approach would be
setting up the vmcs fields through one or several world switch(es) and
then ask tyche to seal the domain, before doing the
driver_switch_domain.

This commit needs further fixups.

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>