x86/coco: add tyche TD0
Tyche TD0 does not implement memory acceptance for now. Unaccepted
memory was added in UEFI 2.9. The TD needs to accept the memoory given
by the VMM. The UEFI firmware needs to communicate which memory regions
has been accepted through a bitmap to the kernel, and allocate_e820
allocates the bitmap if unaccepted memory is present. Ideally, we should
implement this part inside tyche as well so that each TD can choose
whether or not it accepts the memory or not, but we can skip it for now.

Tyche TD0 does not implement memory encryption. Instead, we make sure
the tyche regions between different domains are configured correctly.

Tyche TD0 will forces SWIOTLB for all of the DMA regions. The SWIOTLB
is essentially a bounce buffer and marked as shared (and decrypted) in
the current kernel implementation. Ideally, a shared tyche region
between the host kvm and the TD0 will be mapped into this region.

Tyche TD0 does not support transferring pages between private and
shared memory for now.

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>