dma/swiotlb: tyche guest
When we run the linux guest under tyche, the guest should ask tyche for
the capability of the shared region of the domain first. This part is
implemented using a tyche enumeration call. After we get the host
physical address of the shared region, the memblock allocation directly
takes place on the returned guest physical address. If an alias is found
for the region (not identity-mapped), the current implementation will
opt for the alias start address instead.

Now the bounce buffer is allocated successfully. We use the
memblock_alloc_try_nid function to directly request a bounce buffer on
the specific physical memory range. In the test below, I have split the
physical memory into two regions: one confidential and one shared. The
confidential one has range of [0x0, 0x100000000), while the shared
memory ranges from [0x100000000, 0x1589e0000). The allocation skips the
one on the private region, and allocates the SWIOTLB (64MB) on the
shared & active region.

[    0.000000] software IO TLB: area num 1.
[    0.000000] software IO TLB: swiotlb allocation with tyche guest
[    0.000000] software IO TLB: tyche_enum: next=0
[    0.000000] software IO TLB: enum returns: start=0x0, end=0x100000000, cap_type=0x3f00, alias=0x0, next=0x1
[    0.000000] software IO TLB: region is not shared
[    0.000000] software IO TLB: tyche_enum: next=1
[    0.000000] software IO TLB: enum returns: start=0x100000000, end=0x1589e0000, cap_type=0x3d00, alias=0x0, next=0x2
[    0.000000] software IO TLB: allocating a swiotlb buffer on start=0x100000000, size=67108864
[    0.000000] memblock_alloc_try_nid: 67108864 bytes align=0x1000 nid=-1 from=0x0000000100000000 max_addr=0x0000000104000000 swiotlb_init_remap+0x119/0x420
[    0.000000] memblock_reserve: [0x0000000100000000-0x0000000103ffffff] memblock_alloc_range_nid+0xe1/0x160
[    0.000000] software IO TLB: tlb=0xffff888100000000, __pa(tlb)=0x100000000
[    0.000000] memblock_alloc_try_nid: 786432 bytes align=0x1000 nid=-1 from=0x0000000000000000 max_addr=0x0000000000000000 swiotlb_init_remap+0x18e/0x420
[    0.000000] memblock_reserve: [0x00000001588e4000-0x00000001589a3fff] memblock_alloc_range_nid+0xe1/0x160
[    0.000000] memblock_alloc_try_nid: 16 bytes align=0x40 nid=-1 from=0x0000000000000000 max_addr=0x0000000000000000 swiotlb_init_remap+0x1bd/0x420
[    0.000000] memblock_reserve: [0x00000001589a8000-0x00000001589a800f] memblock_alloc_range_nid+0xe1/0x160
[    0.000000] software IO TLB: tlb=0xffff888100000000, __pa(tlb)=0x100000000

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>