Fixing GP in tyche
The tyche kvm backend was failing due to a GP at address 0xcf60.
After investigation, I realised that the previous instruction, at
0xcf5d, was trying to set bit 0 in cr0, i.e., enter protected mode.

After doing a diff between vanilla KVM and the tyche backend, I realised
that the set cr0 was failing in tyche.
This was due to the fact that the operand, stored in rcx, is read
directly from vcpu.arch.regs, which is not populated in tyche and thus
was 0. As a result, the value written in cr0 was 0 rather than 0x11.

To solved the issue, I implemented a new function in themis/tyche.c that
forwards the General Purpose register reads to tyche and thus reads the
right value.

This might need to change in the future, or similar issues might arise
due to sets rather than read. We might want to capture general purpose
registers and copy them into the destination domain?