dma: move tyche ops out of swiotlb
It seems like the coherent dma operations does not go through swiotlb...
So, I'm going to modify some of the dma allocation functions under
kernel/dma to send over the region to tyche.

Separate out the tyche operation to a separate header file so that we
can call these from other files.

This is temporary as this should eventually be consistent with the
in-kernel tyche driver, so we'll get rid of this commit eventually...

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>