dma: modify IOMMU's mapping on coherent dma allocations
As the coherent dma buffer allocation does not go through swiotlb, we
have to update the IOMMU mapping on the I/O Domain once there's a
coherent dma region allocation.

Note that this patch does not guarantee the allocation will occur on the
tyche's partition of shared memory. The current way to prevent this from
happening is through segmentation. As we have to segment the shared
region and then send the region over to the I/O domain, we'll just
directly fail on segmentataion if we have a region outside of the shared
region. I need to fix this so that the guest will only allocate the
coherent dma area on the shared region on tyche.

Also this patch introduces some really ugly global variables to keep the
handle from tyche's side because I just want to get things working
first... I'll see if there's a way to eliminate these variables

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>