dma/swiotlb: static dma buffer allocation at early boot stage
To have the attestation working, we need a static IOPT mapping on the
IOMMU. Given this requirement, the linux guest needs to be able to
accept a static mapping at boot time and guarantee the allocation on
these dma holes. Moreover, we want to have multiple bounce buffers, so
that each device can have its own dedicated bounce buffer.

The default way to allocate buffer on swiotlb is to use the memblock
allocator in the early boot stage. This works most of the time but it's
not perfect, as the allocation might fail. CMA is not a good solution
either, as it uses the memblock allocator underneath...

To guarantee the allocation, we use the memmap kernel parameter to
directly reserve a physical memory range on the e820 bios. After the
memory allocator is initialized, we use the ioremap() to remap these
reserved memory back into the page table. Note it is important to wait
until the kernel slub allocator is initialized to call the ioremap().
Otherwise, the memblock allocator stage is too early for the ioremap (as
it internally calls vmalloc()).

The code is super messy (because I'm still debugging)... Just a backup
for now

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>