dma/swiotlb: add support for coherent memory allocation
Given that we want a static iopt mapping, we need to make sure the
dma_alloc_coherent actually goes through the swiotlb range. The
restricted DMA buffer patches already provides some insights, but only
works with device trees. This patch adds a separate memory allocator on
the swiotlb range in case when we're running tyche. This memory
allocator is responsible for one range of the shared memory under tyche,
and all of the device's coherent memory allocation will go through this
allocator.

I'll do some cleaning on this patch soon...

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>