dma/swiotlb: tyche TD0 changes
The coherent DMA buffer allocation will be inside the specified swiotlb
range. A separate memory allocator is added for this purpose so that the
coherent DMA buffer allocation will go through functions like
`swiotlb_alloc` and `swiotlb_free`.

We need to turn off the CONFIG_SWIOTLB_DYNAMIC so that linux does not
allocate buffer at runtime.

Tyche needs to first specify some shared regions and the TD0 will first
enumerate the capa and then allocates the bounce buffer on this range.
Once the bounce buffer is allocated successfully, the TD0 asks tyche
to segment the region and will send the capa of the shared region to
tyche. Tyche then updates the IOPT on the IOMMU. There will be no
updates to the IOMMU when the TD0 kernel is running. The TD0 should
not modify the IOMMU and will boot up with "iommu=off".

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>