Linux with embedded tyche driver and loadable KVM themis backend
arch/x86/themis:
This is the tyche kvm-backend that talks to the monitor instead of
configuring Intel VT-x hardware directly.

drivers/tyche:
A tyche driver that provides an API to manage domains.
It is responsible for keeping track of the current domain (TD0)
capabilities and performing the right calls to the monitor.
Some of its functions are exposed to the rest of the kernel (e.g., the
themis backend) via direct calls thanks to EXPORT_SYMBOL.
The driver also has a user-interface via a character device driver.
This is typically used by the sdk-tyche.

drivers/contalloc:
Simple continuous memory allocator.
This is mainly used by the sdk-kvm to replace the mmap interface of the
tyche driver.

Modifications in Linux itself:
We added a flag to kvm.h and changed virt/kvm/kvm_main.c to allow more
than 2 flags to be used in a user mem region flags to encode tyche
access rights and segment types (confidential).