:star: Confidential VMs with devices
This commit provides support for confidential VMs.

In order to work out-of-the-box with virtio device drivers, we leverage
two mechanisms:

1. Software IO TLBs (SWIOTLBs) for packets. This triggers bounce buffers
   when any content is sent to a device (i.e., a copy into a known
   memory range of the packet). The kernel should be configured with
   SWIOTLB support (see configs in tyche-devel/configs) and the kernel
   booted with `swiotlb=force` kernel command line argument.
2. CMA for device virtio queues. Linux supports Contiguous Memory
   Allocation for PCI (virtio) devices virtio queues
   (virtqueues/vrings). Unfortunately, this is skipped if the queue is
   less than or equal to a page. This commit forces the use of CMA
   memory even if the queue is <= PAGE_SIZE if CONFIG_CONFIDENTIAL_VM == 1.
   The correct features must be enabled in the Kconfig for Linux to have
   the expected behavior (see tyche-devel/configs) and the kernel
   command line must request cma (e.g., cma=100k).

The tyche driver and libcapa are modified to support sharing both the
SWIOTLB and CMA buffers when transitioning into confidential mode.
One notable change in the libcapa is the refactoring of the internal
capa grant/share to extract the code that acts directly on capabilities
(and separate it from the one that requires to have a child_domain_t
structure).
The driver relies on the default channel capability to the parent to send the
two shared regions described above.

On the capability side:

Adding support for multi-return value in capabilities
We want to differentiate between regular switch (synch) and async
returns due to exceptions.
This is necessary to run gramine and KVM-sdk